backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
supabase>=2.4.0
sentence-transformers>=3.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0

# Uncomment for LLM_BACKEND=qwen
torch>=2.2.0
transformers>=4.40.0
accelerate>=0.29.0

# Uncomment for LLM_BACKEND=openai or azure
# openai>=1.30.0

redis>=5.0.0